requirements.txt
aiohttp>=3.8.1
numpy>=1.20.0
asyncio>=3.4.3
python-dateutil>=2.8.2
python-dotenv>=1.0.0
requests>=2.31.0
matplotlib>=3.7.0
pandas>=2.0.0
pyyaml>=6.0
